api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.39
alembic==1.15.2
psycopg[binary]==3.2.6
pydantic==2.11.1
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
boto3==1.37.24
email-validator==2.2.0
